requirements.txt
Django~=5.1
djangorestframework~=3.15.2
phonenumbers~=8.13.43
twilio~=9.2.3
